Revert "Moved rep status widget code to platform/user folder (#35559)"
This reverts commit 66164696489b2fe484bd89c0b14a4624776d48bb.
