+++ src/applications/static-pages/representative-status/tests/representativeTypeMap.unit.spec.js
