+++ src/applications/static-pages/representative-status/utilities/formatContactInfo.js
