+++ src/applications/static-pages/representative-status/utilities/representativeTypeMap.js
